Correction d'erreurs mineures Suppresion de commentaire inutile Réagencement des import et des fonctions Renommage du fichier
